Documents/vAIb/quoter/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
requests
jinja2
aiofiles
reportlab 